+++ resources/lang/en/blog.php
@@ -3,6 +3,7 @@
 return [
 
     'blog_empty' => 'Create there your adventures!',
+    'new_blog' => 'New adventure',
     'new_blog_g' => 'New group adventure',
     'adventure_name' => 'Name of the adventure',
     'empty_posts' => 'It seems that you haven\'t writen anything yet. Are you in the mood now?',
@@ -11,5 +12,6 @@ return [
     'content' => 'Content',
     'content_placeholder' => 'Write there great stuff <i class="fad fa-smile-wink colored"></i>',
     'class_blog' => 'This adventure is shared with all the classroom, everyone will read your posts!',
+    'group_blog' => 'This adventure is shared with your team members.',
 
 ];